Add init target to help manage folding
When stowing to a completely bare/new user directory, the lack of any
existing directories means that stow will end up folding all of the
targets that it stows. This is usually not what's wanted, so to help
avoid this situation I created a helper script `.stow-fold.sh` which
will look a the targets to be stowed, and create a directory structure
for them before stowing, thus avoiding unwanted folding. To opt any
directories into folding it looks for the presence of a `.stow-fold`
file, in which case it doesn't create any child directories for that
target.